Provides routes from the router definition file if exists; fallsback to old fs strategy.
We check for the `router.js` file and parse the routes from the file, obeying
`resetNamespace` option on the routes. Since the router.js is a single source
of information about the routes, this would cater to the following types of
projects: mu, pod, classic.

The only caveats are

1. Magic router defintions like , where adminRouter is routes from `ember-admin`

```
   Router.map(function() {
     adminRouter(this);
   });
```

since we list the routes from addons seperately, this should not be a problem.
But if the routes are 'mounted' on a sub route, then we'd loose the context.

2. Addons

We'd use the FS way here, as the router .js would not be present.
So no behavioural change from the previous implementation, but the routes could
be mounted on a sub route as in the previous case.

The only erroneous case will be when the `router.js` file itself has lot of
magic in it - in which case we don't list the routes from both sources. But I
think this situation would be too rate for us to worry about.