log to root of ELS package
This creates a debugging log file at the root of the ELS package, rather than inside the `lib` directory.

The previous location was extremely difficult to find, since it’s alongside the compiled TypeScript output.

An environment variable has also been added to more easily turn on the file logger.